JSON deserializon fix
 * Int attributes changed to double

+++ Challenger/Challenger.Web/Models/Total.cs
@@ -5,13 +5,13 @@ namespace Challenger.Web.Models
     public class Total
     {
         [JsonProperty(PropertyName = "burgers_burned")]
-        public int BurgersBurned { get; set; }
+        public double BurgersBurned { get; set; }
         [JsonProperty(PropertyName = "calories")]
-        public int Calories { get; set; }
+        public double Calories { get; set; }
         [JsonProperty(PropertyName = "distance")]
-        public int Distance { get; set; }
+        public double Distance { get; set; }
         [JsonProperty(PropertyName = "duration")]
-        public int Duration { get; set; }
+        public double Duration { get; set; }
         [JsonProperty(PropertyName = "workouts")]
         public int Workouts { get; set; }
     }
